roachtest: combine the split roachtests
Now that we have more than one roachtests testing different
split logic, it makes sense to have a single unified roachtest
file where all of these are being tested. More will be added here
too.

Release note: None
